back/requirements.txt
requests==2.28.2
idna==3.4
certifi==2022.12.7
urllib3==1.26.14
Django==4.1.6
pytz==2022.7.1
asgiref==3.6.0
sqlparse==0.4.3
djangorestframework==3.14.0
beautifulsoup4==4.11.2
soupsieve==2.3.2
greenlet==2.0.2
typing-extensions==4.4.0
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
anyio==3.6.2
sniffio==1.3.0
starlette==0.22.0
lxml==4.9.2
click==8.1.3
h11==0.14.0
aiohttp==3.8.3
python-dotenv==0.21.1

